utility tests added
