+++ app/inbound/save-invoice-number.js
@@ -1,5 +1,5 @@
 const db = require('../data')
-const reverseEngineerInvoiceNumber = require('../processing/reverse-engineer-invoice-number')
+const { reverseEngineerInvoiceNumber } = require('../utility')
 
 const saveInvoiceNumber = async (invoiceNumber, transaction) => {
   await db.invoiceNumber.findOrCreate({
