--- app/processing/reverse-engineer-invoice-number.js
@@ -1,5 +1,0 @@
-const reverseEngineerInvoiceNumber = (invoiceNumber) => {
-  return `original${invoiceNumber.slice(0, 5)}`
-}
-
-module.exports = reverseEngineerInvoiceNumber
