+++ app/utility/index.js
@@ -1,9 +1,9 @@
 const convertToPence = require('./convert-to-pence')
 const convertToPounds = require('./convert-to-pounds')
-const reverseInvoiceNumber = require('./reverse-invoice-number')
+const reverseEngineerInvoiceNumber = require('./reverse-engineer-invoice-number')
 
 module.exports = {
   convertToPence,
   convertToPounds,
-  reverseInvoiceNumber
+  reverseEngineerInvoiceNumber
 }
