+++ app/utility/reverse-engineer-invoice-number.js
@@ -1,7 +1,7 @@
-const reverseInvoiceNumber = (invoiceNumber) => {
+const reverseEngineerInvoiceNumber = (invoiceNumber) => {
   const firstPart = 'SFI'
   const secondPart = invoiceNumber.slice(1, 8)
   return firstPart.concat(secondPart)
 }
 
-module.exports = reverseInvoiceNumber
+module.exports = reverseEngineerInvoiceNumber
