+++ test/mock-components/mock-invoice-number.js
@@ -2,5 +2,5 @@ const SFI_FIRST_PAYMENT = 'S0000001SFIP000001V001'
 
 module.exports = {
   SFI_FIRST_PAYMENT,
-  SFI_FIRST_PAYMENT_ORIGINAL: SFI_FIRST_PAYMENT.slice(8, 18)
+  SFI_FIRST_PAYMENT_ORIGINAL: 'SFI0000001'
 }
